driver/hid/Keyboardio: Do not compile if not using KeyboardioHID
KeyboardioHID depends on PluggableUSB, which is something completely different
than the USB stack was using prior. Some Arduino cores do not have PluggableUSB,
and as such, can't use KeyboardioHID. So for our KeyboardioHID driver, lets skip
those platforms.

The easiest way to skip them is to check for `USBD_USE_HID_COMPOSITE`, which
signals the core we want to use composite HID.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>