driver::hid: Remove Mouse::getReport()
This function is completely unused, and should not have been exposed to begin
with. Neither in the base class, nor in specific implementations.

The `getReport()` method ties us to a specific report datatype, both in name,
and in shape. That's not something we want, we'd rather have the base class be
HID-library agnostic, which it can't be with `getReport()` present.

Luckily, the function is completely unused, and as such, is safe to remove
without deprecation.

Signed-off-by: Gergely Nagy <algernon@keyboard.io>